dataformat - uint8_t
send gsensor_array_uint8
[x_H][x_L][y_H][y_L][z_H][z_L]

+++ source/getGensor.h
@@ -8,6 +8,6 @@
 #ifndef GETGENSOR_H_
 #define GETGENSOR_H_
 
-int16_t * getGSensor(void);
+uint8_t * getGSensor(void);
 
 #endif /* GETGENSOR_H_ */
